refactoring: utils functions
