+++ src/utils/formatToCurrency.tsx
@@ -1,0 +1,6 @@
+export default function formatToCurrency(value: number | string): string {
+  return new Intl.NumberFormat('pt-BR', {
+    style: 'currency',
+    currency: 'BRL',
+  }).format(Number(value));
+}
